refactor(util): 패키지 구조 변경
- 기존의 baseball 패키지에 있던 ErrorDetector, Input, Output
- util로 이동
