+++ src/main/java/util/ErrorDetector.java
@@ -1,4 +1,4 @@
-package baseball;
+package util;
 
 import constants.MessageConstants;
 import constants.NumberConstants;
