+++ src/main/java/util/Output.java
@@ -1,4 +1,4 @@
-package baseball;
+package util;
 
 import constants.MessageConstants;
